requirements.txt
stremalit
st-gsheets-connection
plotly_express
plotly
pandas
datetime
matplotlib.pyplot